Enhance player spawn logic and testing utilities
- Improved player spawn separation logic in the game reducer to ensure starting units are placed at a minimum distance based on map size.
- Added hex distance calculation to validate spawn distances between players.
- Updated tests to utilize a new test utility for initial game state, ensuring consistency across test cases.
- Created a comprehensive test for spawn distance validation, ensuring players start at least N hexes apart.
- Refactored existing tests to import initial state from the new test utilities module, reducing dependencies on production code.
