+++ package.json
@@ -41,7 +41,9 @@
     "@vitejs/plugin-react": "*",
     "@vitest/browser": "^3.2.4",
     "@vitest/coverage-v8": "*",
-  "zod": "^3.23.0",
+    "playwright":"latest",
+    "@playwright/test":"*",
+    "zod": "^3.23.0",
     "axe-core": "*",
     "eslint": "*",
     "eslint-config-prettier": "*",
