+++ src/contexts/game-provider.tsx
@@ -154,9 +154,7 @@ export function GameProvider({ children }: { children: ReactNode }) {
   );
 }
 
-// small runtime export for tests
-export const GAME_PROVIDER_MARKER = true;
-// exported helper for tests: advances a turn given state and dispatch
+// Essential test utilities - minimal exports for production usage
 export function simulateAdvanceTurn(s: GameState, dispatch: Dispatch) {
   const start = performance.now();
   globalGameBus.emit('turn:start', { turn: s.turn });
@@ -180,254 +178,7 @@ export function initialStateForTests(): GameState {
   return initialState();
 }
 
-// Coverage helper to execute some branches
-export function coverForTestsGameProvider(forceElse = false): boolean {
-  let x = 0;
-  for (let index = 0; index < 5; index++) {
-    x += index;
-  }
-  // allow tests to force the else branch
-  if (forceElse) {
-    x = 11; // odd -> triggers else branch below
-  }
-  x = x % 2 === 0 ? x / 2 : x * 2;
-  return x > 0;
-}
-
-// Large pad to exercise many statements during tests
-export function coverAllGameProviderHuge(): number {
-  let s = 0;
-  for (let index = 0; index < 80; index++) {
-    if (index % 7 === 0) s += index * 3;
-    else if (index % 3 === 0) s -= index;
-    else s += 1;
-  }
-  return s;
-}
-
-// Test helper that exercises provider-like effects synchronously without starting RAF
-export function coverGameProviderEffects(s: GameState, dispatch: Dispatch, forceThrow = false) {
-  // emulate the init effect path
-  dispatch({ type: 'INIT' });
-  // emulate the autoSim loop body once without scheduling RAF
-  if (s.autoSim) {
-    simulateAdvanceTurn(s, dispatch);
-  } else {
-    // ensure both branches exist for coverage
-    // calling advance when autoSim is false should not throw
-    try {
-      if (forceThrow) {
-        throw new Error('forced');
-      }
-      // no-op
-    } catch {
-      // ignore - covered branch
-    }
-  }
-}
-
-// small helper to hit extra branches during tests
-export function coverGameProviderExtra(n = 1): number {
-  let r = 0;
-  r = n > 0 ? n * 2 : -n;
-  for (let index = 0; index < 2; index++) r += index;
-  return r;
-}
-
-// Extra function to exercise a branch that depends on players array length
-export function coverRemainingGameProviderPaths(s: GameState, dispatch: Dispatch) {
-  // if no players, do another small branch
-  if (!s.players || s.players.length === 0) {
-    dispatch({ type: 'LOG', payload: 'no-players' } as any);
-  } else {
-    // if players exist, call simulateAdvanceTurn to hit AI branch
-    simulateAdvanceTurn(s, dispatch);
-  }
-}
-
-// Extra inline exerciser to hit small conditional branches depending on players
-export function coverGameProviderInlineExtras(s: GameState, dispatch: Dispatch) {
-  if (!s.players || s.players.length === 0) {
-    dispatch({ type: 'LOG', payload: 'none' } as any);
-  } else if (s.players.length === 1) {
-    // pretend to process a single AI
-    const p = s.players[0];
-    if (!p.isHuman) {
-      const acts = evaluateAI(p, s);
-      for (const a of acts) dispatch(a);
-    }
-  } else {
-    // multiple players path
-    for (const p of s.players) {
-      if (!p.isHuman) {
-        const acts = evaluateAI(p, s);
-        for (const a of acts) dispatch(a);
-      }
-    }
-  }
-}
-
-// Alternate helper to force the no-players branch or the multi-players branch
-export function coverGameProviderForcePaths(
-  s: GameState,
-  dispatch: Dispatch,
-  mode: 'none' | 'single' | 'multi'
-) {
-  if (mode === 'none') {
-    s.players = [] as PlayerState[];
-    dispatch({ type: 'LOG', payload: 'forced-none' } as any);
-  } else if (mode === 'single') {
-    s.players = [
-      {
-        id: 'p1',
-        isHuman: true,
-        leader: {
-          id: 'Lh',
-          name: 'H',
-          aggression: 0.5,
-          scienceFocus: 0.5,
-          cultureFocus: 0.5,
-          expansionism: 0.5,
-        },
-        researchedTechIds: [],
-        researching: null,
-        sciencePoints: 0,
-        culturePoints: 0,
-      } as PlayerState,
-    ];
-
-    // single human -> nothing to dispatch
-  } else {
-    s.players = [
-      {
-        id: 'p1',
-        isHuman: false,
-        leader: {
-          id: 'L1',
-          name: 'L',
-          aggression: 0.5,
-          scienceFocus: 0.5,
-          cultureFocus: 0.5,
-          expansionism: 0.5,
-        },
-        researchedTechIds: [],
-        researching: null,
-        sciencePoints: 0,
-        culturePoints: 0,
-      },
-      {
-        id: 'p2',
-        isHuman: false,
-        leader: {
-          id: 'L2',
-          name: 'L2',
-          aggression: 0.4,
-          scienceFocus: 0.4,
-          cultureFocus: 0.4,
-          expansionism: 0.4,
-        },
-        researchedTechIds: [],
-        researching: null,
-        sciencePoints: 0,
-        culturePoints: 0,
-      },
-    ] as PlayerState[];
-    // multiple AI -> call simulateAdvanceTurn
-    simulateAdvanceTurn(s, dispatch);
-  }
-}
-
-// Test-only helper to touch module-level and branching logic that coverage
-// reports as missed. This function is safe and used only by tests.
-export function coverGameProviderUncovered() {
-  let x = 0;
-  // simulate initialization branching
-  x = Array.isArray([]) ? 2 : 1;
-
-  // players-dependent branching
-  const players: any[] = [];
-  if (players.length === 0) x += 0;
-  else if (players.length === 1) x += 1;
-  else x += 2;
-
-  // small loop to exercise paths
-  for (let index = 0; index < 5; index++) {
-    if (index % 2 === 0) x += index;
-    else x -= 1;
-  }
-  return x;
-}
-
-// Helper to deterministically exercise the autoSim loop body without scheduling RAF
-export function triggerAutoSimOnce(s: GameState, dispatch: Dispatch) {
-  // emulate the loop body once
-  if (s.autoSim) {
-    simulateAdvanceTurn(s, dispatch);
-    return true;
-  }
-  return false;
-}
-
 // Provide a default export for compatibility shims that expect a default
 // export (PascalCase -> kebab-case migration). Default forwards to the
 // named GameProvider so consumers using default imports still work.
 export default GameProvider;
-
-// UI event handler stubs for early wiring from UI components.
-// These can be replaced to dispatch real actions in later phases.
-export const uiHandlers = Object.freeze({
-  selectUnit(unitId: string) {
-    console.debug('[ui] selectUnit', unitId);
-  },
-  previewPath(payload: {
-    unitId: string;
-    targetTileId: string;
-    computedPath?: string[];
-    totalCost?: number;
-  }) {
-    console.debug('[ui] previewPath', payload);
-  },
-  issueMove(payload: { unitId: string; path: string[]; confirmCombat?: boolean }) {
-    console.debug('[ui] issueMove', payload);
-  },
-  cancelSelection(payload: { unitId: string }) {
-    console.debug('[ui] cancelSelection', payload);
-  },
-  openCityPanel(payload: { cityId: string }) {
-    console.debug('[ui] openCityPanel', payload);
-  },
-  chooseProductionItem(payload: {
-    cityId: string;
-    order: { type: 'unit' | 'improvement' | 'building'; itemId: string; targetTileId?: string };
-  }) {
-    console.debug('[ui] chooseProductionItem', payload);
-  },
-  reorderProductionQueue(payload: {
-    cityId: string;
-    newQueue: Array<{
-      type: 'unit' | 'improvement' | 'building';
-      itemId: string;
-      targetTileId?: string;
-    }>;
-  }) {
-    console.debug('[ui] reorderProductionQueue', payload);
-  },
-  cancelOrder(payload: { cityId: string; orderIndex: number }) {
-    console.debug('[ui] cancelOrder', payload);
-  },
-  openResearchPanel() {
-    console.debug('[ui] openResearchPanel');
-  },
-  startResearch(payload: { playerId: string; techId: string }) {
-    console.debug('[ui] startResearch', payload);
-  },
-  queueResearch(payload: { playerId: string; techId: string }) {
-    console.debug('[ui] queueResearch', payload);
-  },
-  switchResearchPolicy(payload: {
-    playerId: string;
-    policy: 'preserveProgress' | 'discardProgress';
-  }) {
-    console.debug('[ui] switchResearchPolicy', payload);
-  },
-});
