+++ tests/gameprovider_autosim_true.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { coverGameProviderEffects, initialStateForTests } from '../src/contexts/game-provider';
+import { coverGameProviderEffects, initialStateForTests } from '../src/test-utils/game-provider';
 
 describe('GameProvider autoSim branch', () => {
   it('runs simulateAdvanceTurn when s.autoSim is true', () => {
