+++ tests/gameprovider_final.test.ts
@@ -3,7 +3,7 @@ import {
   initialStateForTests,
   simulateAdvanceTurn,
   coverGameProviderEffects,
-} from '../src/contexts/game-provider';
+} from '../src/test-utils/game-provider';
 import { applyAction } from '../src/game/reducer';
 import { globalGameBus } from '../src/game/events';
