+++ tests/save_file.test.ts
@@ -1,6 +1,6 @@
-import { vi } from 'vitest';
+import { describe, test, expect, vi } from 'vitest';
 import { exportToFile, importFromFile, serializeState, SizeExceededError } from '../src/game/save';
-import { initialStateForTests } from '../src/contexts/game-provider';
+import { initialStateForTests } from '../src/test-utils/game-provider';
 
 describe('file save helpers', () => {
   test('exportToFile triggers download', () => {
