+++ tests/save_roundtrip.test.ts
@@ -1,10 +1,11 @@
+import { describe, test, expect } from 'vitest';
 import {
   serializeState,
   deserializeState,
   VersionMismatchError,
   ValidationError,
 } from '../src/game/save';
-import { initialStateForTests } from '../src/contexts/game-provider';
+import { initialStateForTests } from '../src/test-utils/game-provider';
 
 describe('save/load roundtrip', () => {
   test('roundtrip retains state', () => {
