+++ tests/unit-spawn-detailed.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { applyAction } from '../src/game/reducer';
-import { initialStateForTests } from '../src/contexts/game-provider';
+import { initialStateForTests } from '../src/test-utils/game-provider';
 import { BiomeType } from '../src/game/types';
 
 describe('Unit Spawn Terrain Detailed Testing', () => {
