+++ tests/unitMovement_combatGuard.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { applyAction } from '../src/game/reducer';
-import { initialState } from '../src/contexts/game-provider';
+import { initialStateForTests as initialState } from '../src/test-utils/game-provider';
 
 describe('issueMove combat guard', () => {
   it('blocks entering enemy tile unless confirmCombat is true', () => {
